requirements.txt
python-fasthtml
monsterui
PyYAML
fastcore
numpy